+++ dust-lang/src/instruction/address.rs
@@ -10,37 +10,37 @@ use super::MemoryKind;
     Clone, Copy, Debug, Default, Eq, PartialEq, PartialOrd, Ord, Hash, Serialize, Deserialize,
 )]
 pub struct Address {
-    pub index: usize,
+    pub index: u16,
     pub memory: MemoryKind,
 }
 
 impl Address {
-    pub fn new(index: usize, memory: MemoryKind) -> Self {
+    pub fn new(index: u16, memory: MemoryKind) -> Self {
         Self { index, memory }
     }
 
-    pub fn cell(index: usize) -> Self {
+    pub fn cell(index: u16) -> Self {
         Address {
             index,
             memory: MemoryKind::CELL,
         }
     }
 
-    pub fn constant(index: usize) -> Self {
+    pub fn constant(index: u16) -> Self {
         Address {
             index,
             memory: MemoryKind::CONSTANT,
         }
     }
 
-    pub fn register(index: usize) -> Self {
+    pub fn register(index: u16) -> Self {
         Address {
             index,
             memory: MemoryKind::REGISTER,
         }
     }
 
-    pub fn encoded(index: usize) -> Self {
+    pub fn encoded(index: u16) -> Self {
         Address {
             index,
             memory: MemoryKind::ENCODED,
@@ -49,14 +49,14 @@ impl Address {
 
     pub fn function_self() -> Self {
         Address {
-            index: usize::MAX,
+            index: u16::MAX,
             memory: MemoryKind::CONSTANT,
         }
     }
 
     pub fn display(&self, f: &mut Formatter<'_>, r#type: OperandType) -> fmt::Result {
         if r#type == OperandType::FUNCTION {
-            if self.index == usize::MAX {
+            if self.index == u16::MAX {
                 write!(f, "self")
             } else {
                 write!(f, "proto_{}", self.index)
